refactor: reorganize page imports and update routing paths for payment orders, accounts, suppliers, and details
