+++ src/pages/details/manage_details.py
